requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv
